+++ Sources/TuistLoader/Models+ManifestMappers/Config+ManifestMapper.swift
@@ -13,7 +13,9 @@ extension TuistGraph.Config {
     ///   - path: The path of the config file.
     static func from(manifest: ProjectDescription.Config, at path: AbsolutePath) throws -> TuistGraph.Config {
         let generatorPaths = GeneratorPaths(manifestDirectory: path)
-        let generationOptions = try manifest.generationOptions.map { try TuistGraph.Config.GenerationOption.from(manifest: $0, generatorPaths: generatorPaths) }
+        let generationOptions = try manifest.generationOptions.map {
+            try TuistGraph.Config.GenerationOption.from(manifest: $0, generatorPaths: generatorPaths)
+        }
         let compatibleXcodeVersions = TuistGraph.CompatibleXcodeVersions.from(manifest: manifest.compatibleXcodeVersions)
         let plugins = try manifest.plugins.map { try PluginLocation.from(manifest: $0, generatorPaths: generatorPaths) }
         let swiftVersion: TSCUtility.Version?
@@ -59,7 +61,10 @@ extension TuistGraph.Config.GenerationOption {
     /// - Parameters:
     ///   - manifest: Manifest representation of Tuist config generation options
     ///   - generatorPaths: Generator paths.
-    static func from(manifest: ProjectDescription.Config.GenerationOptions, generatorPaths: GeneratorPaths) throws -> TuistGraph.Config.GenerationOption {
+    static func from(
+        manifest: ProjectDescription.Config.GenerationOptions,
+        generatorPaths: GeneratorPaths
+    ) throws -> TuistGraph.Config.GenerationOption {
         switch manifest {
         case let .xcodeProjectName(templateString):
             return .xcodeProjectName(templateString.description)